requirements.txt
flask
psycopg2-binary
psycopg2
bcrypt
pyjwt
flask-jwt-extended
flask-sqlalchemy
flask-migrate
flask-jwt-extended
pika
flask-restx